moved updated sync classes to new sync.py
