+++ dsi/sync.py
@@ -17,7 +17,7 @@ class Sync():
     """
     A class defined to assist in data management activities for DSI
 
-    Sync is where data movement functions such as copy (to remote location) and 
+    Sync is where data movement functions such as copy (to remote location) and
     sync (local filesystem with remote) exist.
     """
     def __init__(self, project_name="test"):
@@ -29,12 +29,15 @@ class Sync():
                 extension = ext
                 break
         if extension != "":
-            f = self.project_name + extension
+            self.full_db_name = self.project_name + extension
         else:
-            f = self.project_name+".db"
-            for ext in (".duckdb", ".sqlite", ".db", ".sqlite3"):
+            proj_db_found = False
+            for ext in (".db", ".duckdb", ".sqlite", ".sqlite3"):
                 if os.path.exists(self.project_name + ext):
-                    f = self.project_name + ext
+                    if proj_db_found:
+                        raise ValueError(f"Multiple databases found with {project_name}. Specify an extension.")
+                    self.full_db_name = self.project_name + ext
+                    proj_db_found = True
 
         self.remote_location = None
         self.local_location = None
@@ -50,35 +53,28 @@ class Sync():
         if create_bool is False:
             raise RuntimeError(f"Cannot open the {project_name} database due to write permissions. Please try elsewhere.")
     
-        backend_name = self.t.identify_backend(f)
+        backend_name = self.t.identify_backend(self.full_db_name)
         if backend_name is None:
             raise ValueError("Unsupported DSI database type. Currently supporting: Sqlite, DuckDB")
 
         fnull = open(os.devnull, 'w')
         with redirect_stdout(fnull):
-            self.t.load_module('backend', backend_name, 'back-write', filename=f)
+            self.t.load_module('backend', backend_name, 'back-write', filename=self.full_db_name)
 
         if not self.t.valid_backend(self.t.loaded_backends[0], self.t.loaded_backends[0].__class__.__bases__[0].__name__):
             raise RuntimeError(f"{project_name} database must have metadata in it before trying to call DSI move functions.")
 
-    def execute_cmd(self, cmd, cmd_name):
-        process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True, encoding='latin-1')
-        stdout, stderr = process.communicate()
-        if process.returncode != 0:
-            raise RuntimeError(f"{cmd_name} failed: {stderr}")
-        return stdout
-    
     def reindex(self, local_loc, remote_loc, isVerbose = False):
         """
-        Helper function that allows users to index their data again by dropping existing fileystem information.
+        Helper function that allows users to index their data again by dropping existing filesystem information.
         """
-        # drop filesystem table and call index()
-        # future -- use existing db to "reindex" by updating the filepath cols, not dropping table.
+        # current -- drop filesystem table and call index()
+        # future --- use existing db to "reindex" by updating the filepath cols, not dropping table.
         #   remove filesystem pass through in query_artifacts if not dropping table
         self.t.artifact_handler(interaction_type='query', query = "DROP TABLE IF EXISTS filesystem;")
         self.index(local_loc, remote_loc, isVerbose)
     
-    def index(self, local_loc, remote_loc, isVerbose=False):
+    def index(self, local_loc, remote_loc, isVerbose=False, no_parent = False):
         """
         Helper function to gather filesystem information, local and remote locations
         to create a filesystem entry in a new or existing database
@@ -91,11 +87,11 @@ class Sync():
             if isVerbose:
                 print("Warning: filesystem table already exists! DSI Index skipped.")
             return
-        
+
         # Relative paths (..) will not work
-        if ".." in local_loc:
+        if "../" in local_loc or "../" in remote_loc:
             raise ValueError("Error: Please use absolute paths instead of relative")
-        
+
         if isVerbose:
             print("loc: "+local_loc+ " rem: "+remote_loc)
 
@@ -106,7 +102,9 @@ class Sync():
             file_list, tmp = itertools.tee(file_list)
             file_len=sum(1 for _ in tmp)
             print("Crawled "+str(file_len)+" files.")
-
+        
+        file_list = list(file_list) # save as list since dircrawl2() returns an iterator 
+        
         self.remote_location = remote_loc
         self.local_location = local_loc
         # populate st_list to hold all filesystem attributes
@@ -136,11 +134,15 @@ class Sync():
             last = -10
 
         for file in file_list:
+            parent_rel_file = Path(file).relative_to(Path(local_loc).parent)
             rel_file = os.path.relpath(file,local_loc) #rel path
             filepath = os.path.join(local_loc, rel_file)
             st = os.stat(filepath)
             # append future location to st
-            rfilepath = os.path.join(remote_loc,self.project_name, rel_file)
+            if no_parent: # exclude parent dir of every file in remote location
+                rfilepath = os.path.join(remote_loc,self.project_name, rel_file)
+            else:
+                rfilepath = os.path.join(remote_loc,self.project_name, parent_rel_file)
             rfile_list.append(rfilepath)
             st_dict['file_origin'].append(rel_file)
             st_dict['size'].append(st.st_size)
@@ -165,29 +167,12 @@ class Sync():
 
         if isVerbose:
             print(f"] Collection object created with {len(st_list)} entries.")
-
-        # Test remote location validity, try to check access
-        # Future: iterate through remote/server list here, for now:::
-        remote_list = [ os.path.join(remote_loc,self.project_name) ]
-        for remote in remote_list:
-            if isVerbose:
-                print(f"Testing access to '{remote}' directory.")
-            try: # Try for file permissions
-                if os.path.exists(remote): # Check if exists
-                    print(f"The directory '{remote}' already exists remotely.")
-                else:
-                    path = Path(remote)
-                    path.mkdir(parents=True, exist_ok=True)
-                    # os.makedirs(remote) # Create it
-                    print(f"The directory '{remote}' has been created remotely.")
-            except Exception as err:
-                raise RuntimeError(f"Error creating remote directory: {err}")
                 
         if isVerbose:
             print("Creating filesystem table")
         fnull = open(os.devnull, 'w')
         with redirect_stdout(fnull):
-            self.t.load_module('plugin', "Dict", "reader", collection=st_dict, table_name="filesystem")
+            self.t.load_module('plugin', "Dictionary", "reader", collection=st_dict, table_name="filesystem")
             self.t.artifact_handler(interaction_type='ingest')
 
         self.file_list = file_list
@@ -199,13 +184,53 @@ class Sync():
     def move(self, tool="copy", isVerbose=False, **kwargs):
         self.copy(tool,isVerbose,kwargs)
 
+    def execute_cmd(self, cmd, cmd_name, timer = False):
+        """Internal helper for Sync to call executable actions"""
+        process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True, encoding='latin-1')
+
+        if timer:
+            start = time.time()
+            while process.poll() is None:
+                elapsed = int(time.time() - start)
+                print(f"\rRunning... {elapsed}s elapsed", end="", flush=True)
+                time.sleep(10)
+
+        stdout, stderr = process.communicate()
+        if process.returncode != 0:
+            if "too many authentication failures" in str(stderr).lower():
+                raise RuntimeError(f"{cmd_name} failed due to multiple incorrect password attempts. Check the password and remote path.")
+            raise RuntimeError(f"{cmd_name} failed: \n{stderr}")
+        return stdout
+    
+    def change_group(self, local_loc, user_group):
+        """Change group permissions for data and db. Only works for OS with Unix (not Windows)"""
+        try:
+            cmd = ["chgrp", "-R", user_group, local_loc]
+            self.execute_cmd(cmd, "changing user group for data", True)
+
+            cmd = ["chgrp", user_group, self.full_db_name]
+            self.execute_cmd(cmd, "changing user group for database")
+        except Exception as e:
+            print("Warning:", str(e))
+
+    def change_permissions(self, local_loc):
+        """Change read permissions for data and db. Only works for OS with Unix (not Windows)"""
+        try:
+            cmd = ["chmod", "-R", "750", local_loc] # 770 to make read/write to all. 750 to make read to all
+            self.execute_cmd(cmd, "changing read permissions for data", True)
+
+            cmd = ["chmod", "750", self.full_db_name]
+            self.execute_cmd(cmd, "changing read permissions for database")
+        except Exception as e:
+            print("Warning:", str(e))
+    
     def copy(self, tool="copy", isVerbose=False, **kwargs):
         """
         Helper function to perform the data copy over using a preferred API
         """
         if any(x is None for x in (self.remote_location, self.local_location, self.file_list, self.rfile_list)):
             raise RuntimeError("Must run successful DSI Index right before Copy")
-        
+
         fnull = open(os.devnull, 'w')
         with redirect_stdout(fnull):
             fs_t = self.t.get_table("filesystem")
@@ -213,9 +238,27 @@ class Sync():
             print(" filesystem table not found. Must run Index first.")
             print(" Data copy failed.")
             return
-
-        # Future: have movement service handle type (cp,scp,ftp,rsync,etc.)
-        if tool == "copy":
+        
+        # Test remote location validity and try creating folders
+        # Future: iterate through remote/server list here, for now:
+        if tool.lower() not in ["scp", "rsync"]: # Exclude scp and rsync since they create folders differently
+            remote_list = [ os.path.join(self.remote_location,self.project_name) ]
+            for remote in remote_list:
+                if isVerbose:
+                    print(f"Testing access to '{remote}' directory.")
+                try: # Try for file permissions
+                    if os.path.exists(remote): # Check if exists
+                        print(f"The directory '{remote}' already exists remotely.")
+                    else:
+                        path = Path(remote)
+                        path.mkdir(parents=True, exist_ok=True)
+                        # os.makedirs(remote) # Create it
+                        print(f"The directory '{remote}' has been created remotely.")
+                except Exception as err:
+                    raise RuntimeError(f"Error creating remote directory: {err}")
+        
+        # Future: have movement service handle type without user input (cp,scp,ftp,rsync,etc.)
+        if tool.lower() == "copy":
             # Data movement via Unix Copy
             for file,file_remote in zip(self.file_list,self.rfile_list):
                 abspath = os.path.dirname(os.path.abspath(file_remote))
@@ -227,24 +270,95 @@ class Sync():
                         path.mkdir(parents=True)
                     except Exception:
                         raise RuntimeError(f"Unable to create folder {abspath}. Check your access rights")
-            
+
                 if isVerbose:
                     print(" cp " + file + " " + file_remote)
                 shutil.copy2(file , file_remote)
-                
+
             # Database movement
             if isVerbose:
-                print(" cp " + str(self.project_name+".db") + " " + os.path.join(self.remote_location, self.project_name, self.project_name+".db"))
-            shutil.copy2(str(self.project_name+".db"), os.path.join(self.remote_location, self.project_name, self.project_name+".db"))
+                print(" cp " + self.full_db_name + " " + os.path.join(self.remote_location, self.project_name, self.full_db_name))
+            shutil.copy2(self.full_db_name, os.path.join(self.remote_location, self.project_name, self.full_db_name))
 
             print(" Data Copy Complete!")
         
-        elif tool == "scp":
-            #  Data movement via SCP
-            remote_user = os.getlogin()
-            remote_host = "myremote"
+        elif tool.lower() == "scp":
+            try:
+                host_part, path_part = self.remote_location.split(":", 1)
+            except ValueError:
+                raise ValueError("Remote path must be in the format user@host:/absolute/path")
+
+            if not path_part.startswith("/") and "nt" not in os.name:
+                raise ValueError("Remote path must be absolute (starting with /)")
+            
+            # making remote dir
+            if isVerbose:
+                print(" ssh "+ str(host_part) + " \"mkdir -p " + str(os.path.join(path_part, self.project_name)) + "\"" )
+            cmd = ["ssh", host_part, f'mkdir -p \"{os.path.join(path_part, self.project_name)}\"']
+            print("Creating remote directory if it doesn't exist")
+            self.execute_cmd(cmd, "Creating remote dir")
+
+
+            #remove username from file_remote column in filesystem table
+            username, host = host_part.split("@")
+            filesystem_df = self.t.get_table("filesystem")
+            filesystem_df["file_remote"] = filesystem_df["file_remote"].str.replace(f"{username}@", "", regex=False)
+
+            self.t.dsi_tables.remove("filesystem")
+            self.t.overwrite_table("filesystem", filesystem_df)
+            self.t.dsi_tables.append("filesystem")
+
+            cmd = ["scp", "-rp", self.local_location, os.path.join(self.remote_location, self.project_name)]
+            if isVerbose:
+                print()
+                print(*cmd)
+            self.execute_cmd(cmd, "scp data")
+            print(" DSI SCP data movement complete.")
+
+            cmd = ["scp", "-p", self.full_db_name, os.path.join(self.remote_location, self.project_name, self.full_db_name)]
+            if isVerbose:
+                print()
+                print(*cmd)
+            self.execute_cmd(cmd, "scp database")
+            print(" DSI SCP database movement complete.")
         
-        elif tool == "conduit":
+        elif tool.lower() == "rsync":
+            try:
+                host_part, path_part = self.remote_location.split(":", 1)
+            except Exception:
+                raise ValueError("Remote location must be in the format user@host:/absolute/path")
+
+            if not path_part.startswith("/"):
+                raise ValueError("Remote path must be absolute (starting with /)")
+            
+            #remove username from file_remote column in filesystem table
+            try:
+                username, host = host_part.split("@")
+            except Exception:
+                raise ValueError("Remote path's hostname must be in the format user@server") from None
+            filesystem_df = self.t.get_table("filesystem")
+            filesystem_df["file_remote"] = filesystem_df["file_remote"].str.replace(f"{username}@", "", regex=False)
+
+            self.t.dsi_tables.remove("filesystem")
+            self.t.overwrite_table("filesystem", filesystem_df)
+            self.t.dsi_tables.append("filesystem")
+            
+            self.local_location = self.local_location[:-1] if self.local_location.endswith("/") else self.local_location
+            cmd = ["rsync", "-av", f"--rsync-path=mkdir -p {os.path.join(path_part, self.project_name)} && rsync", 
+                   self.local_location, os.path.join(self.remote_location, self.project_name)]
+            if isVerbose:
+                print(*cmd)
+            self.execute_cmd(cmd, "rsync data")
+            print(" DSI Rsync data movement complete.")
+            
+            cmd = ["rsync", "-av", self.full_db_name, os.path.join(self.remote_location, self.project_name)]
+            if isVerbose:
+                print()
+                print(*cmd)
+            self.execute_cmd(cmd, "rsync database")
+            print(" DSI Rsync database movement comlpete.")
+        
+        elif tool.lower() == "conduit":
             import signal
 
             # Test Kerberos
@@ -255,7 +369,7 @@ class Sync():
             if "No credentials" in stdout:
                 print("Kerberos authentication error: No credentials found. Please type 'conduit get' to reissue a ticket.")
                 raise RuntimeError("Kerberos message: " + str(stdout))
-            
+
             # Test Conduit status
             def alarm_handler(signum, frame):
                 raise RuntimeError("Conduit not authenticated. Please type 'conduit get' to issue a ticket.")
@@ -267,14 +381,14 @@ class Sync():
                     print("Testing Conduit: conduit get")
                 cmd = ['/usr/projects/systems/conduit/bin/conduit-cmd','--config','/usr/projects/systems/conduit/conf/conduit-cmd-config.yaml','get']
                 stdout = self.execute_cmd(cmd, "Testing conduit get")
-                
+
                 if "TRANSFER_ID" in stdout and isVerbose:
                     print(" Conduit is authenticated.")
                 elif "TRANSFER_ID" not in stdout:
                     raise RuntimeError("Conduit Error: " + str(stdout))
             finally:
                 signal.alarm(0)
-            
+
             try:
                 base_cmd = ['/usr/projects/systems/conduit/bin/conduit-cmd','--config','/usr/projects/systems/conduit/conf/conduit-cmd-config.yaml','cp','-r']
                 # File Movement
@@ -286,20 +400,19 @@ class Sync():
 
                 # Database Movement
                 if isVerbose:
-                    print("conduit cp " + str(self.project_name+".db") + " " + os.path.join(self.remote_location, self.project_name, self.project_name+".db"))
-                cmd = base_cmd + [str(self.project_name+".db"), os.path.join(self.remote_location, self.project_name, self.project_name+".db")]
+                    print("conduit cp " + self.full_db_name + " " + os.path.join(self.remote_location, self.project_name, self.full_db_name))
+                cmd = base_cmd + [self.full_db_name, os.path.join(self.remote_location, self.project_name, self.full_db_name)]
                 self.execute_cmd(cmd, "Conduit copy database")
                 print(" DSI submitted Conduit database movement job.")
 
                 print("Type 'conduit get' to track status of both jobs.")
-                print("  If 'WaitingForLease' status, that is fine.")
-                print("  If 'Error' status, type 'conduit error <TRANSFER_ID>' to view detailed error ouput.")
+                print("  If 'WaitingForLease' status, data move is in queue.")
+                print("  If 'Error' status, type 'conduit error <TRANSFER_ID>' to view detailed error output.")
 
-            except subprocess.CalledProcessError as e:
-                raise RuntimeError(f"Conduit failed with error: {e.stderr} ")
+            except Exception as e:
+                raise RuntimeError(f"Conduit failed with error: {str(e)} ")
 
-
-        elif tool == "pfcp":           
+        elif tool.lower() == "pfcp":           
             try:
                 # File Movement
                 if isVerbose:
@@ -310,16 +423,16 @@ class Sync():
 
                 # Database Movement
                 if isVerbose:
-                    print("pfcp " + str(self.project_name+".db") + " " + os.path.join(self.remote_location, self.project_name, self.project_name+".db"))
-                cmd = ['pfcp', str(self.project_name+".db"), os.path.join(self.remote_location, self.project_name, self.project_name+".db")]
+                    print("pfcp " + self.full_db_name + " " + os.path.join(self.remote_location, self.project_name, self.full_db_name))
+                cmd = ['pfcp', self.full_db_name, os.path.join(self.remote_location, self.project_name, self.full_db_name)]
                 self.execute_cmd(cmd, "pfcp move database")
                 print(" DSI submitted pfcp database movement job.")
-            except subprocess.CalledProcessError as e:
-                print(f"Command failed with error: {e.stderr} ")
+            except Exception as e:
+                raise RuntimeError(f"pfcp failed with error: {str(e)} ")
         
-        elif tool == "ftp":
+        elif tool.lower() == "ftp":
             True
-        elif tool == "git":
+        elif tool.lower() == "git":
             True
         else:
             raise TypeError(f"Data movement format not supported:, Type: {tool}")
@@ -334,7 +447,7 @@ class Sync():
         `return`: returns crawled file-list
         """
         start_time = time.perf_counter()
-        
+
         file_list = []
         for root, dirs, files in os.walk(filepath):
             #if os.path.basename(filepath) != 'tmp': # Lets skip some files
@@ -343,18 +456,18 @@ class Sync():
                 print(f"Crawling directory: {root}")
                 print(f"  Found {len(files)} files, {len(dirs)} subdirectories")
 
-            for f in files: # Appent root-level files
+            for f in files: # Append root-level files
                 file_list.append(os.path.join(root, f))
-            
+
         elapsed = time.perf_counter() - start_time
 
         if verbose:
             print(f"\nFinished crawling: {filepath}")
             print(f"Total files found: {len(file_list)}")
             print(f"Runtime: {elapsed:.2f} seconds")
-    
+
         return file_list
- 
+
     def dircrawl2(self, filepath: str, verbose: bool = False) -> Iterator[str]:
         start = time.perf_counter()
 
@@ -390,7 +503,7 @@ class Sync():
     def gen_uuid(self, st):
         '''
         Generates a unique file identifier using the os.stat data object as the input
-        
+
         '''
         inode=st.st_ino
         ctime=st.st_ctime
@@ -408,14 +521,14 @@ class TarFile():
     self.local_tmp_dir = local_tmp_dir
     self.local_files = local_files
     self.create_tar(self.local_files)
-      
+
   def create_tar(self, local_files=[]):
     """
     Creates a tar file and returns the index
-      
+
     tar_name: name of the tar file to create with .tar.gz as the extension
     local_files: a list of files with full paths to include
-    
+
     The tar file will be created in the local_tmp_dir directory
     """
 
@@ -424,7 +537,7 @@ class TarFile():
             os.mkdir(self.local_tmp_dir)
         except Exception as err:
             print(f"Unexpected {err=}, {type(err)=}")
-    
+
     self.tar_path = self.local_tmp_dir + "/" + self.tar_name
     tar = tarfile.open(self.tar_path, "w:gz")
     for f in local_files:
@@ -448,18 +561,18 @@ class TarFile():
 
   def get_name(self):
       return self.tar_name
-  
+
 class HPSSSync():
     """
     A class defined to assist in HPSS data management activities for DSI
 
-    Sync is where data movement functions such as copy (to HPSS) and 
+    Sync is where data movement functions such as copy (to HPSS) and
     sync (local filesystem with remote) exist.
     """
     remote_dir = None
     local_files = []
     tar_files = []
-    
+
     def __init__(self, project_name="test"):
         # Helper function to get parent module names.
 
@@ -467,16 +580,16 @@ class HPSSSync():
 
     def index(self, local_files, remote_dir, tar_name, isVerbose=False):
         """
-        Helper function to gather local file information and to create a tar that should 
+        Helper function to gather local file information and to create a tar that should
         be in remote_dir or will be copied to the remote directory
 
         local_files: a list of files or directories to add to a tar
         remote_dir: a directory on HPSS that has or will have the tar file
         """
-        
+
         if isVerbose:
             print("loc: "+ str(local_files) + " hpss remote: "+remote_dir)
-        
+
         # Tar the local_files list
         tar_file = TarFile(tar_name, local_files)
         self.local_files = local_files
@@ -531,7 +644,7 @@ class HPSSSync():
             with redirect_stdout(fnull):
                 t.load_module('backend','Sqlite','back-read', filename=f)
         except Exception:
-            print(f"Databaase {f} not found")
+            print(f"Database {f} not found")
             raise
 
         # See if filesystem exists
@@ -554,11 +667,11 @@ class HPSSSync():
                 writer = csv.DictWriter(csvfile, fieldnames=st_dict.keys())
                 writer.writeheader()
                 writer.writerows(rows)
-            
+
             # Add filesystem table
             t.load_module('plugin', 'Csv', 'reader', filenames=".fs.csv", table_name="filesystem_hpss")
             t.artifact_handler(interaction_type='ingest')
-        
+
         t.close()
 
     def move(self, tool="copy", isVerbose=False, **kwargs):
@@ -566,7 +679,7 @@ class HPSSSync():
 
     def copy(self, tool="copy", isVerbose=False, **kwargs):
         """
-        Helper function to perform the 
+        Helper function to perform the
         data copy over using a preferred API
         """
         # See if FS table has been created
@@ -581,7 +694,7 @@ class HPSSSync():
             fnull = open(os.devnull, 'w')
             with redirect_stdout(fnull):
                 t.load_module('backend','Sqlite','back-read', filename=f)
-        except Exception as err:
+        except Exception:
             print(f"Database {f} not found")
             raise
 
@@ -591,9 +704,9 @@ class HPSSSync():
             print( " Filesystem table not found. Try running Index first.")
             print( " Data copy failed. ")
             return
-        
+
         t.close()
-      
+
         hpss_files = {}
         for f in self.tar_files:
             hpss_files[self.remote_dir + "/" + f.get_name()] = f.get_full_path()
